+++ clang/lib/Sema/SemaDecl.cpp
@@ -13449,6 +13449,21 @@ void Sema::AddInitializerToDecl(Decl *RealDecl, Expr *Init, bool DirectInit) {
     IsParenListInit = !InitSeq.steps().empty() &&
                       InitSeq.step_begin()->Kind ==
                           InitializationSequence::SK_ParenthesizedListInit;
+    QualType VDeclType = VDecl->getType();
+    if (!VDeclType->isDependentType() &&
+        Context.getAsIncompleteArrayType(VDeclType)) {
+      if (Init && !Init->getType().isNull() &&
+          !Init->getType()->isDependentType()) {
+        if (Context.getAsIncompleteArrayType(Init->getType())) {
+          // Bail out if it is not possible to deduce array size from the
+          // initializer.
+          Diag(VDecl->getLocation(), diag::err_typecheck_decl_incomplete_type)
+              << VDeclType;
+          VDecl->setInvalidDecl();
+          return;
+        }
+      }
+    }
   }
 
   // Check for self-references within variable initializers.